--renamed command/common to command/utils
